fix: review: integrate configuration
