--- piikii-output-cache/redis/src/main/kotlin/com/piikii/output/redis/config/RedisCacheConfig.kt
@@ -1,35 +1,0 @@
-package com.piikii.output.redis.config
-
-import org.springframework.cache.CacheManager
-import org.springframework.cache.annotation.EnableCaching
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.redis.cache.RedisCacheConfiguration
-import org.springframework.data.redis.cache.RedisCacheManager
-import org.springframework.data.redis.connection.RedisConnectionFactory
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer
-import org.springframework.data.redis.serializer.RedisSerializationContext
-import org.springframework.data.redis.serializer.StringRedisSerializer
-import java.time.Duration
-
-@Configuration
-@EnableCaching
-class RedisCacheConfig {
-    @Bean
-    fun cacheManager(redisConnectionFactory: RedisConnectionFactory): CacheManager {
-        val redisCacheConfiguration =
-            RedisCacheConfiguration.defaultCacheConfig()
-                .entryTtl(Duration.ofDays(7))
-                .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(StringRedisSerializer()))
-                .serializeValuesWith(
-                    RedisSerializationContext.SerializationPair.fromSerializer(
-                        GenericJackson2JsonRedisSerializer(),
-                    ),
-                )
-
-        return RedisCacheManager.RedisCacheManagerBuilder
-            .fromConnectionFactory(redisConnectionFactory)
-            .cacheDefaults(redisCacheConfiguration)
-            .build()
-    }
-}
